test(frontend): mockProjectEventStream 回落本地，createDeferred 恢复可选 reason
CONTRIBUTING 的共享设施标准把 API spy 组合排除在收编之外，且该工厂只有
useProjectEventsSSE.test.tsx 一个消费者，回落为该文件的本地辅助；替身仍取
共享 FakeEventSource。createDeferred 的 reject 恢复 origin/main 中
useAssistantSession.test.tsx 原实现的 `reason?: unknown` 签名。
