+++ frontend/src/test/deferred.ts
@@ -4,7 +4,7 @@
  */
 export function createDeferred<T>() {
   let resolve!: (value: T) => void;
-  let reject!: (reason: unknown) => void;
+  let reject!: (reason?: unknown) => void;
   const promise = new Promise<T>((res, rej) => {
     resolve = res;
     reject = rej;
